Consolidated code between stix_validator.py and cybox_validator.py scripts. * Added an __init__.py to scripts subpackage. This is where the   common code lives between the two scripts. * Added a profile_to_xslt() and profile_to_schematron() methods   to sdv package. * Added is_cybox() and is_stix() checks to validate() calls on   STIX and CybOX XML schema validator classes.
