+++ examples/best_practice_invalid.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 <stix:STIX_Package
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xmlns:stix="http://stix.mitre.org/stix-1"
