+++ sdv/validators/stix/schema.py
@@ -70,6 +70,13 @@ class STIXSchemaValidator(object):
                 "method"
             )
 
+    def _check_root(self, doc):
+        if utils.is_stix(doc):
+            return
+
+        error = "Input document does not contain a valid STIX root element."
+        raise errors.ValidationError(error)
+
     def validate(self, doc, version=None, schemaloc=False):
         """Performs XML Schema validation against a STIX document.
 
@@ -100,6 +107,9 @@ class STIXSchemaValidator(object):
 
         """
         root = utils.get_etree_root(doc)
+
+        # check that this is a STIX document
+        self._check_root(root)
 
         if schemaloc:
             validator = self._xml_validators[self._KEY_SCHEMALOC]
